# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Booking Management API
servers:
  - url: /api
    description: Base API path
...
paths:
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all property settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update one or more settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all rooms
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: roomId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: propertyId
          in: query
          required: false
          schema:
            type: integer
        - name: unitId
          in: query
          required: false
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
...
  /work-orders:
    get:
      operationId: listWorkOrders
      tags: [workOrders]
      summary: List work orders
      parameters:
        - name: propertyId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [workOrders]
      summary: Create a work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      parameters:
        - name: propertyId
          in: query
          required: false
          schema:
            type: integer
        - name: role
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications (optionally unread only)
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppNotification"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppNotification"
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts
      parameters:
        - name: propertyId
          in: query
          required: false
          schema:
            type: integer
        - name: staffId
          in: query
          required: false
          schema:
            type: integer
        - name: startDate
          in: query
          required: false
          schema:
            type: string
          description: "YYYY-MM-DD"
        - name: endDate
          in: query
          required: false
          schema:
            type: string
          description: "YYYY-MM-DD"
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create a shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
...
  /guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List all unique guests with their booking history
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name, email, or phone
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
...
  /users:
    get:
      operationId: listUsers
      summary: List all users
      tags: [users]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PmsUser"
    post:
      operationId: createUser
      summary: Create a user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmsUser"
...
  /activity-logs:
    get:
      operationId: listActivityLogs
      summary: List activity audit logs
      tags: [activity-logs]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: entityType
          in: query
          schema:
            type: string
        - name: actorRole
          in: query
          schema:
            type: string
        - name: propertyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
  /maintenance-requests:
    post:
      operationId: createMaintenanceRequest
      summary: Submit a maintenance request (guest or staff)
      tags: [maintenance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceRequestInput"
      responses:
        "201":
          description: Work order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /field-users:
    get:
      operationId: listFieldUsers
      tags: [fieldUsers]
      summary: List field operations users
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of field users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FieldUser"
    post:
      operationId: createFieldUser
      tags: [fieldUsers]
      summary: Create a field user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldUserInput"
      responses:
        "201":
          description: Field user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldUser"
...
components:
  schemas:
    Settings:
      type: object
      required: [propertyName, logoText, logoSub, businessMode, enabledModules, companyName, contactEmail, contactPhone, contactAddress, taskTypes, taskRequirements]
      properties:
        propertyName:
          type: string
          description: Display name shown as the dashboard heading and sidebar logo
        logoText:
          type: string
          description: Large serif word in the sidebar logo
        logoSub:
          type: string
          description: Smaller sans-serif word in the sidebar logo
        businessMode:
          type: string
          description: "hotel | compound | tower | serviced-apartments"
        enabledModules:
          type: array
          items:
            type: string
          description: List of enabled functional module IDs for this client
        companyName:
          type: string
          description: Legal or trading name of the company
        contactEmail:
          type: string
          description: Primary contact email address
        contactPhone:
          type: string
          description: Primary contact phone number
        contactAddress:
          type: string
          description: Office / property address
        taskTypes:
          type: array
          items:
            $ref: "#/components/schemas/TaskType"
          description: Custom task category definitions
        taskRequirements:
          $ref: "#/components/schemas/TaskRequirements"
    SettingsInput:
      type: object
      properties:
        propertyName:
          type: string
        logoText:
          type: string
        logoSub:
          type: string
        businessMode:
          type: string
        enabledModules:
          type: array
          items:
            type: string
        companyName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        contactAddress:
          type: string
        taskTypes:
          type: array
          items:
            $ref: "#/components/schemas/TaskType"
        taskRequirements:
          $ref: "#/components/schemas/TaskRequirements"
...
    Room:
      type: object
      required: [id, name, type, pricePerNight, status, createdAt]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "e.g. Standard, Deluxe, Suite, Presidential"
        pricePerNight:
          type: number
        status:
          type: string
          description: "available, occupied, maintenance"
        description:
          type: ["string", "null"]
        capacity:
          type: integer
        amenities:
          type: ["string", "null"]
        createdAt:
          type: string
    RoomInput:
      type: object
      required: [name, type, pricePerNight, status]
      properties:
        name:
          type: string
        type:
          type: string
        pricePerNight:
          type: number
        status:
          type: string
        description:
          type: string
        capacity:
          type: integer
        amenities:
          type: string
...
    Booking:
      type: object
      required: [id, guestName, guestEmail, roomId, checkIn, checkOut, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: ["string", "null"]
        roomId:
          type: integer
        roomName:
          type: ["string", "null"]
        roomType:
          type: ["string", "null"]
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
          description: "confirmed, checked-in, checked-out, cancelled"
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [guestName, guestEmail, roomId, checkIn, checkOut, totalAmount]
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        roomId:
          type: integer
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        notes:
          type: string
...
    Property:
      type: object
      required: [id, name, address, city, country, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          description: "active, inactive"
        unitCount:
          type: integer
        createdAt:
          type: string
    PropertyInput:
      type: object
      required: [name, address, city]
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        description:
          type: string
        status:
          type: string
...
    Expense:
      type: object
      required: [id, propertyId, title, category, amount, expenseDate, createdAt]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        unitId:
          type: ["integer", "null"]
        unitName:
          type: ["string", "null"]
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [propertyId, title, category, amount, expenseDate]
      properties:
        propertyId:
          type: integer
        unitId:
          type: integer
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
        notes:
          type: string
...
    WorkOrder:
      type: object
      required: [id, propertyId, title, priority, status, createdAt]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        unitId:
          type: ["integer", "null"]
        unitName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          description: "low, medium, high, urgent"
        status:
          type: string
          description: "pending, in-progress, on-hold, completed"
        assignedTo:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    WorkOrderInput:
      type: object
      required: [propertyId, title, priority]
      properties:
        propertyId:
          type: integer
        unitId:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        assignedTo:
          type: string
        cost:
          type: number
        dueDate:
          type: string
...
    StaffMember:
      type: object
      required: [id, name, role, email, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        propertyId:
          type: ["integer", "null"]
        propertyName:
          type: ["string", "null"]
        status:
          type: string
          description: "active, inactive"
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, role, email]
      properties:
        name:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyId:
          type: integer
        status:
          type: string
...
    AppNotification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          description: "check-in, check-out, maintenance-alert"
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        relatedId:
          type: ["integer", "null"]
        relatedType:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Guest:
      type: object
      required: [guestName, guestEmail, totalBookings, totalSpent, lastStay, firstStay]
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: ["string", "null"]
        totalBookings:
          type: integer
        totalSpent:
          type: number
        lastStay:
          type: ["string", "null"]
        firstStay:
          type: ["string", "null"]
        lastRoomType:
          type: ["string", "null"]
        lastStatus:
          type: ["string", "null"]
...
    Shift:
      type: object
      required: [id, staffId, propertyId, date, shiftType, startTime, endTime, createdAt]
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        staffRole:
          type: ["string", "null"]
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        date:
          type: string
        shiftType:
          type: string
          description: "morning, afternoon, evening, night"
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ShiftInput:
      type: object
      required: [staffId, propertyId, date, shiftType, startTime, endTime]
      properties:
        staffId:
          type: integer
        propertyId:
          type: integer
        date:
          type: string
        shiftType:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
...
    PmsUser:
      type: object
      required: [id, username, displayName, role, permissions, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, displayName, password, role]
      properties:
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...
    ActivityLog:
      type: object
      required: [id, action, entityType, createdAt]
      properties:
        id:
          type: integer
        actorId:
          type: integer
          nullable: true
        actorName:
          type: string
          nullable: true
        actorRole:
          type: string
          nullable: true
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
          nullable: true
        entityLabel:
          type: string
          nullable: true
        propertyId:
          type: integer
          nullable: true
        propertyName:
          type: string
          nullable: true
        details:
          type: string
          nullable: true
        assignedByName:
          type: string
          nullable: true
        completedByName:
          type: string
          nullable: true
        proofPhotoUrl:
          type: string
          nullable: true
        createdAt:
          type: string
...